working on how things will process. have a constant offset file added, and changed the dyn file a bit. also hueristics better sorted

+++ common/kdlsym.cpp
@@ -17,7 +17,7 @@
     #define kernel_symbol_t kernel_symbol
 #endif
 
-int ksym_dlsym_kcrc(kernel_linux* kernel_local_target, const char* newString, size_t* out_address)
+int ksym_dlsym_kcrc(kernel_linux* kernel_local_target, const char* newString, uint64_t* out_address)
 {
     int result = -1;
     const char* kstrBase = 0;
